Change project structure to multi-module
