+++ bot/build.gradle.kts
@@ -1,0 +1,15 @@
+dependencies {
+    implementation(project(":kryo-discord-whitelist-common"))
+
+    implementation("net.dv8tion:JDA:5.0.0-alpha.3") {
+        exclude(module = "opus-java")
+    }
+
+    tasks {
+        withType<Jar> {
+            manifest {
+                attributes["Main-Class"] = "org.kryonite.kryodiscordwhitelist.bot.KryoDiscordWhitelistBot"
+            }
+        }
+    }
+}
