+++ bot/src/main/resources/logback.xml
