+++ common/build.gradle.kts
@@ -1,0 +1,3 @@
+dependencies {
+    implementation("org.mariadb.jdbc:mariadb-java-client:2.7.4")
+}
