+++ settings.gradle.kts
@@ -1,2 +1,9 @@
-rootProject.name = "kryo-discord-bot"
+rootProject.name = "kryo-discord-whitelist"
 
+include("velocity")
+include("bot")
+include("common")
+
+project(":velocity").name = "kryo-discord-whitelist-velocity"
+project(":bot").name = "kryo-discord-whitelist-bot"
+project(":common").name = "kryo-discord-whitelist-common"
