--- src/main/java/org/kryonite/kryodiscordbot/persistence/entity/User.java
@@ -1,10 +1,0 @@
-package org.kryonite.kryodiscordbot.persistence.entity;
-
-import lombok.Data;
-
-@Data(staticConstructor = "create")
-public class User {
-
-  private final long discordId;
-  private final String minecraftName;
-}
